refactor(app): the Streamlit app is the draft board only (#91)
The picks pages reached parity in the React app some time ago, so keeping a
second implementation of make-picks, view-picks and manage-spreads meant two
places to fix every pool-scoring bug.

Deletes app/pages/view_picks.py, app/pages/manage_spreads.py and the old
app/main.py (make picks). The draft board moves up from app/pages/ to
app/main.py, which makes this a single-page app rather than a multipage one,
and carries run_app() over so the `app` console script still works.

Picks, spreads and standings are now only in web/ (MakePicks, ManageSpreads,
Standings, Field).
